# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  FsGetInfo:
    description: GlueFS의 상세정보 구조체
    properties:
      id:
        type: integer
      mdsmap:
        properties:
          bal_rank_mask:
            type: string
          balancer:
            type: string
          created:
            type: string
          damaged:
            items: {}
            type: array
          data_pools:
            items:
              type: integer
            type: array
          enabled:
            type: boolean
          epoch:
            type: integer
          ever_allowed_features:
            type: integer
          explicitly_allowed_features:
            type: integer
          failed:
            items: {}
            type: array
          flags:
            type: integer
          flags_state:
            properties:
              allow_multimds_snaps:
                type: boolean
              allow_snaps:
                type: boolean
              allow_standby_replay:
                type: boolean
              joinable:
                type: boolean
              refuse_client_session:
                type: boolean
            type: object
          fs_name:
            type: string
          in:
            items:
              type: integer
            type: array
          info: {}
          last_failure:
            type: integer
          last_failure_osd_epoch:
            type: integer
          max_file_size:
            type: integer
          max_mds:
            type: integer
          metadata_pool:
            type: integer
          modified:
            type: string
          required_client_features:
            type: object
          root:
            type: integer
          session_autoclose:
            type: integer
          session_timeout:
            type: integer
          standby_count_wanted:
            type: integer
          stopped:
            items: {}
            type: array
          tableserver:
            type: integer
          up:
            properties:
              mds_0:
                type: integer
            type: object
        type: object
    type: object
  FsStatus:
    description: GlueFS의 상태 구조체
    properties:
      clients:
        items:
          properties:
            clients:
              type: integer
            fs:
              type: string
          type: object
        type: array
      mds_version:
        items:
          properties:
            daemon:
              items:
                type: string
              type: array
            version:
              type: string
          type: object
        type: array
      mdsmap:
        items:
          properties:
            name:
              type: string
            state:
              type: string
          type: object
        type: array
      pools:
        items:
          properties:
            avail:
              type: integer
            id:
              type: integer
            name:
              type: string
            type:
              type: string
            used:
              type: integer
          type: object
        type: array
    type: object
...
  HTTP400BadRequest:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  HTTP404NotFound:
    properties:
      code:
        example: 404
        type: integer
      message:
        example: Not Found
        type: string
    type: object
  HTTP500InternalServerError:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: InternalServerError
        type: string
    type: object
...
info:
  contact:
    email: support@ablecloud.io
    name: 윤여천
    url: http://www.ablecloud.io
  description: This is a GlueAPI server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Glue-API
  version: v1.0
paths:
...
  /api/v1/gluefs:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: GlueFS의 상태값과 리스트를 보여줍니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/FsStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Show Status and List of Glue FS
      tags:
      - GlueFS
    put:
      consumes:
      - application/x-www-form-urlencoded
      description: GlueFS를 수정합니다.
      parameters:
      - description: Glue FS Old Name
        in: formData
        name: old_name
        required: true
        type: string
      - description: Glue FS New Name
        in: formData
        name: new_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Update of Glue FS
      tags:
      - GlueFS
  /api/v1/gluefs/{fs_name}:
    delete:
      consumes:
      - application/x-www-form-urlencoded
      description: GlueFS를 삭제합니다.
      parameters:
      - description: Glue FS Name
        in: path
        name: fs_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Delete of Glue FS
      tags:
      - GlueFS
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: GlueFS를 생성합니다.
      parameters:
      - description: Glue FS Name
        in: path
        name: fs_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Create of Glue FS
      tags:
      - GlueFS
  /api/v1/gluefs/info/{fs_name}:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: GlueFS의 상세 정보를 보여줍니다.
      parameters:
      - description: Glue FS Name
        in: path
        name: fs_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/FsGetInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Detail Info of Glue FS
      tags:
      - GlueFS
...
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"